backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]      # includes pydantic + email-validator
python-multipart     # needed for form/file uploads
